Refactor TimestampExtensions to extensions package and add reactor-netty dependency
