+++ solo-leveling-gateway-service/pom.xml
@@ -93,6 +93,12 @@
       <version>${bcprov-jdk15on.version}</version>
     </dependency>
 
+    <!-- rabbitmq stomp -->
+    <dependency>
+      <groupId>io.projectreactor.netty</groupId>
+      <artifactId>reactor-netty</artifactId>
+    </dependency>
+
     <!-- jwt -->
     <dependency>
       <groupId>io.jsonwebtoken</groupId>
