+++ solo-leveling-gateway-service/src/main/kotlin/com.sleepkqq.sololeveling/gateway/extensions/TimestampExtensions.kt
@@ -1,4 +1,4 @@
-package com.sleepkqq.sololeveling.gateway.exception
+package com.sleepkqq.sololeveling.gateway.extensions
 
 import com.google.protobuf.Timestamp
 import java.time.LocalDateTime
